# crates/bunker-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LAN BUNKER API",
    "description": "The LAN BUNKER backend over SQLite: players and auth, tournaments and brackets, events and check-ins, and the cycles ledger",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/checkin/{code}": {
      "get": {
        "tags": [
          "events"
        ],
        "operationId": "checkin_gate",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CheckinCode"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The event behind the code, and whether its doors are open",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckinGate"
                }
              }
            }
          },
          "400": {
            "description": "A path parameter is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Unknown code, or a draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "events"
        ],
        "operationId": "check_in",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CheckinCode"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Already in. The time of the first scan, and zero cycles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckinReceipt"
                }
              }
            }
          },
          "201": {
            "description": "Checked in, cycles paid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckinReceipt"
                }
              }
            }
          },
          "400": {
            "description": "A path parameter is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No valid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "`PasswordChangeRequired`: the caller must choose a new password first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Unknown code, or a draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "The doors are not open",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/events": {
      "get": {
        "tags": [
          "events"
        ],
        "operationId": "list_events",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The latest night first. Drafts are hidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginated_Event"
                }
              }
            }
          },
          "400": {
            "description": "A query parameter is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/me/checkins": {
      "get": {
        "tags": [
          "events"
        ],
        "operationId": "checkins",
        "responses": {
          "200": {
            "description": "The events the caller checked in to",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Checkins"
                }
              }
            }
          },
          "401": {
            "description": "No valid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "`PasswordChangeRequired`: the caller must choose a new password first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "type": "object",
        "description": "The one shape every failure has on the wire.",
        "required": [
          "code",
          "message",
          "status"
        ],
        "properties": {
          "cause": {
            "type": [
              "string",
              "null"
            ],
            "description": "The chain of causes. It is absent if a user can see the response."
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "CheckinCode": {
        "type": "string",
        "maxLength": 12,
        "minLength": 12,
        "pattern": "^[a-z0-9]{12}$"
      },
      "CheckinGate": {
        "type": "object",
        "description": "Answer of `GET /api/checkin/{code}`: what the door shows before a player\nconfirms. The server decides the window, so no client reads a clock.",
        "required": [
          "event",
          "window"
        ],
        "properties": {
          "event": {
            "$ref": "#/components/schemas/Event"
          },
          "window": {
            "$ref": "#/components/schemas/CheckinWindow"
          }
        }
      },
      "CheckinReceipt": {
        "type": "object",
        "description": "Answer of `POST /api/checkin/{code}`. A repeat answers the time of the\nfirst check-in and pays nothing.",
        "required": [
          "event",
          "checkedInAt",
          "cycles"
        ],
        "properties": {
          "checkedInAt": {
            "type": "string",
            "format": "date-time"
          },
          "cycles": {
            "type": "integer",
            "format": "int64",
            "description": "What this call paid: the check-in cycles the first time, zero after."
          },
          "event": {
            "$ref": "#/components/schemas/Event"
          }
        }
      },
      "CheckinWindow": {
        "type": "string",
        "description": "Where `now` sits against the window of an event. The door pays only while\nit is `Open`.",
        "enum": [
          "early",
          "open",
          "over"
        ]
      },
      "Checkins": {
        "type": "object",
        "description": "Answer of `GET /api/me/checkins`: the events the caller checked in to.",
        "required": [
          "events"
        ],
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventId"
            }
          }
        }
      },
...
      "Description": {
        "type": "string",
        "maxLength": 1000,
        "minLength": 0
      },
...
      "ErrorCode": {
        "type": "string",
        "description": "Each error code the API can answer with. A client selects on these names, so\nthey are part of the contract.\n\nTwo steps make a response from an error: `ServiceError::public` below selects\nthe code, and `From<ErrorCode> for StatusCode` in `server.rs` selects the\nstatus.",
        "enum": [
          "GenericError",
          "ServiceUnavailable",
          "ItemNotFound",
          "HandleTaken",
          "InvalidCredentials",
          "Unauthorized",
          "Forbidden",
          "PasswordChangeRequired",
          "WrongPassword",
          "RegistrationClosed",
          "CheckinClosed",
          "InvalidState",
          "NotAnEntrant",
          "InvalidRequest",
          "UnprocessableRequest",
          "UnsupportedMediaType",
          "PayloadTooLarge",
          "RouteNotFound",
          "MethodNotAllowed"
        ]
      },
      "Event": {
        "type": "object",
        "description": "An event as every client sees it. The check-in code is not here: only an\nadmin reads it, through [`EventDetail`].",
        "required": [
          "id",
          "name",
          "location",
          "games",
          "description",
          "startsAt",
          "endsAt",
          "status",
          "checkinCount",
          "createdAt"
        ],
        "properties": {
          "checkinCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "$ref": "#/components/schemas/Description"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time",
            "description": "The last game."
          },
          "games": {
            "$ref": "#/components/schemas/Games"
          },
          "id": {
            "$ref": "#/components/schemas/EventId"
          },
          "image": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ImageName"
              }
            ]
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "name": {
            "$ref": "#/components/schemas/EventName"
          },
          "startsAt": {
            "type": "string",
            "format": "date-time",
            "description": "The doors open."
          },
          "status": {
            "$ref": "#/components/schemas/EventStatus"
          }
        }
      },
...
      "EventId": {
        "type": "string",
        "format": "uuid"
      },
      "EventName": {
        "type": "string",
        "maxLength": 60,
        "minLength": 1
      },
      "EventStatus": {
        "type": "string",
        "description": "`Draft` is visible to admins only, and its code opens nothing.",
        "enum": [
          "draft",
          "published"
        ]
      },
...
      "Games": {
        "type": "string",
        "maxLength": 200,
        "minLength": 0
      },
...
      "ImageName": {
        "type": "string",
        "description": "A file name under the event images of the site",
        "maxLength": 80,
        "minLength": 1,
        "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$"
      },
...
      "Location": {
        "type": "string",
        "maxLength": 60,
        "minLength": 0
      },
...
      "Paginated_Event": {
        "type": "object",
        "description": "One page of results, and the totals a client needs for a pager.",
        "required": [
          "items",
          "total",
          "page",
          "pageSize",
          "totalPages"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "An event as every client sees it. The check-in code is not here: only an\nadmin reads it, through [`EventDetail`].",
              "required": [
                "id",
                "name",
                "location",
                "games",
                "description",
                "startsAt",
                "endsAt",
                "status",
                "checkinCount",
                "createdAt"
              ],
              "properties": {
                "checkinCount": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "description": {
                  "$ref": "#/components/schemas/Description"
                },
                "endsAt": {
                  "type": "string",
                  "format": "date-time",
                  "description": "The last game."
                },
                "games": {
                  "$ref": "#/components/schemas/Games"
                },
                "id": {
                  "$ref": "#/components/schemas/EventId"
                },
                "image": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/ImageName"
                    }
                  ]
                },
                "location": {
                  "$ref": "#/components/schemas/Location"
                },
                "name": {
                  "$ref": "#/components/schemas/EventName"
                },
                "startsAt": {
                  "type": "string",
                  "format": "date-time",
                  "description": "The doors open."
                },
                "status": {
                  "$ref": "#/components/schemas/EventStatus"
                }
              }
            }
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "maximum": 100,
            "minimum": 1
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "Every row that matches, ignoring the window.",
            "minimum": 0
          },
          "totalPages": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}